aukcija: ispravljen bug prilikom prikazivanja voznji za koje moze da se konkurise

+++ src/Gui/FormeZaDodavanjeIIzmenu/AukcijaVozacForma.java
@@ -44,10 +44,10 @@ public class AukcijaVozacForma extends JFrame {
             long vozacId = vozac.getId();
             long voznjaId = voznja.getId();
             long aukcijaId = Aukcija.generisiIdAukcije();
-            double brojVoznji = Vozac.brojVoznji(vozac);
+            int brojVoznji = Vozac.brojVoznji(vozac);
             double vremeDolaska = Double.parseDouble(txtMinutaDoDestinacije.getText().trim());
             Vozilo vozilo = Vozilo.pretraziPoId(Vozilo.ucitajSvaVozila(),vozac.getBrTaksiVozila());
-            double godinaProizvodnjeVozila = vozilo.getGodProizvodnje();
+            int godinaProizvodnjeVozila = vozilo.getGodProizvodnje();
             double ocenaVozaca = 0;
             double ocena = Aukcija.izracunajOcenu(brojVoznji, vremeDolaska, godinaProizvodnjeVozila, ocenaVozaca);
             Aukcija.sacuvajAukciju(new Aukcija(aukcijaId, vozacId, voznjaId, minutaDoDestinacije, ocena));
